requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Environment
python-dotenv==1.0.0

# AI Integration
google-generativeai==0.3.2

# Email
aiosmtplib==3.0.1

# Background Tasks
APScheduler==3.10.4